requirements.txt
appnope @ file:///Users/runner/miniforge3/conda-bld/appnope_1610094662606/work
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
bleach==3.3.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1618380339289/work
defusedxml==0.7.1
entrypoints==0.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
gspread==3.7.0
idna==2.10
ipykernel @ file:///Users/runner/miniforge3/conda-bld/ipykernel_1617137105563/work/dist/ipykernel-5.5.3-py3-none-any.whl
ipython @ file:///Users/runner/miniforge3/conda-bld/ipython_1616836936704/work
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi @ file:///Users/runner/miniforge3/conda-bld/jedi_1610146802570/work
Jinja2==2.11.3
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1615693636836/work
jupyter-console==6.4.0
jupyter-core @ file:///Users/runner/miniforge3/conda-bld/jupyter_core_1612125274649/work
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
MarkupSafe==1.1.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.3.0
numpy @ file:///Users/runner/miniforge3/conda-bld/numpy_1616933195014/work
oauthlib==3.1.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1617148930513/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
prometheus-client==0.10.1
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1616432837031/work
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1615243893546/work
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1612179539967/work
pyzmq @ file:///Users/runner/miniforge3/conda-bld/pyzmq_1614611820252/work
qtconsole==5.0.3
QtPy==1.9.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
Send2Trash==1.5.0
six @ file:///home/conda/feedstock_root/build_artifacts/six_1590081179328/work
spotipy==2.18.0
terminado==0.9.4
testpath==0.4.4
tornado @ file:///Users/runner/miniforge3/conda-bld/tornado_1610094701146/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1602771532708/work
urllib3==1.26.4
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
widgetsnbextension==3.5.1
